requirements.txt
aiofiles==0.7.0
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.6.0
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
fastapi==0.70.0
fastapi-socketio==0.0.9
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.2
Markdown==3.3.4
MarkupSafe==2.1.1
packaging==21.3
pip==22.1.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.10.2
pyparsing==3.0.9
pyserial==3.5
pytest==6.2.5
python-dotenv==0.19.1
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.1
requests==2.26.0
setuptools==63.4.1
six==1.16.0
sniffio==1.3.0
starlette==0.16.0
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.15.0
wheel==0.37.1
wincertstore==0.2
